% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{auxiliary}
\alias{auxiliary}
\title{auxiliary function}
\usage{
auxiliary(
  callfunction,
  data = NULL,
  opt = NULL,
  param = NULL,
  pred.reps = NULL,
  heteroModel = "BC",
  calc_rho = T,
  meantype = NULL,
  metrics = NULL,
  dir.loc = NULL,
  type.label = NULL,
  box.colour = NULL,
  std.resids = NULL,
  pred.pl = NULL,
  input = NULL
)
}
\arguments{
\item{callfunction}{# specifies which function to be called}

\item{data}{# input data}

\item{opt}{# list of options}

\item{param}{# parameters}

\item{pred.reps}{# replicates}

\item{heteroModel}{# unused}

\item{calc_rho}{# unused}

\item{meantype}{# zero, fitted or linear}

\item{metrics}{# metrics}

\item{dir.loc}{# location of data}

\item{type.label}{# labelling}

\item{box.colour}{# colour of box in boxplot}

\item{std.resids}{# standardised residuals}

\item{pred.pl}{# probability limits}

\item{input}{# axis ranges}
}
\description{
For internal use with the shiny app only. By J Hunter & Team.
}
